src/unittest/python/data_processor_tests.py
# from unittest import TestCase
# from unittest.mock import Mock
# from src.main.python.data_processor import DataProcessor
# from src.main.python.reader_factory import ReaderFactory
# from src.main.python.writer_factory import WriterFactory
#
#
# class TestDataProcessor(TestCase):
#     def test_process_data_with_valid_source_and_destination(self):
#         source_type = ReaderFactory().create_data_source("xml")
#         destination_type = WriterFactory().create_data_destination("sqlite")
#         data_processor = DataProcessor()
#
#         with self.assertLogs() as log:
#             data_processor.process_data(source_type, destination_type, "source_file_path", "destination_file_path")
#
#         self.assertIn("Data processed successfully", log.output[0])
#
#     def test_process_data_with_missing_source(self):
#         data_processor = DataProcessor()
#         with self.assertLogs() as log:
#             data_processor.process_data(None, Mock(), "source_file_path", "destination_file_path")
#         self.assertIn("Invalid source or destination type.", log.output[0])
#
#     def test_process_data_with_missing_destination(self):
#         data_processor = DataProcessor()
#         with self.assertLogs() as log:
#             data_processor.process_data(Mock(), None, "source_file_path", "destination_file_path")
#         self.assertIn("Invalid source or destination type.", log.output[0])
